refactor(card): rm index.js from paths
