+++ packages/card/src/css/index.ts
@@ -11,7 +11,7 @@ import {
 import { names as themeNames } from '@pluralsight/ps-design-system-theme'
 import { transparentize } from '@pluralsight/ps-design-system-util'
 
-import * as vars from '../vars/index.js'
+import * as vars from '../vars'
 
 export default {
   '.psds-card': {
