+++ packages/card/src/index.ts
@@ -1,4 +1,4 @@
-export { default } from './react/index.js'
-export * from './vars/index.js'
-export * from './js/index.js'
-export { default as css } from './css/index.js'
+export { default } from './react'
+export * from './vars'
+export * from './js'
+export { default as css } from './css'
